+++ neoforge/src/main/java/dev/aika/smsn/neoforge/SMSNPlatformImpl.java
@@ -1,5 +1,6 @@
 package dev.aika.smsn.neoforge;
 
+import net.neoforged.fml.loading.FMLConfig;
 import net.neoforged.fml.loading.FMLPaths;
 
 import java.nio.file.Path;
@@ -8,5 +9,13 @@ import java.nio.file.Path;
 public class SMSNPlatformImpl {
     public static Path getConfigDir() {
         return FMLPaths.CONFIGDIR.get();
+    }
+
+    public static boolean isModVersionCheckEnabled() {
+        return FMLConfig.getBoolConfigValue(FMLConfig.ConfigValue.VERSION_CHECK);
+    }
+
+    public static void setModVersionCheckEnabled(boolean enabled) {
+        FMLConfig.updateConfig(FMLConfig.ConfigValue.VERSION_CHECK, enabled);
     }
 }