+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+### Unreleased
+
+**NeoForge**
+
+- Add FML Version Check Warning Screen
+
+**Fabric**
+
+- Add ModMenu Update Checker Warning Screen
+
 ### 1.3.5
 
 **NeoForge**